All RBuilder child declarations are extracted to separate file
