+++ kotlin-react/src/main/kotlin/react/RBuilder.kt
@@ -25,25 +25,6 @@ interface RBuilder {
         childList.add(this)
     }
 
-    fun <P : RProps> child(
-        type: Any,
-        props: P,
-        children: List<Any>
-    ): ReactElement =
-        child(createElement(type, props, *children.toTypedArray()))
-
-    fun <P : RProps> child(
-        type: Any,
-        props: P,
-        handler: RHandler<P>
-    ): ReactElement {
-        val children = with(RElementBuilder(props)) {
-            handler()
-            childList
-        }
-        return child(type, props, children)
-    }
-
     operator fun <P : RProps> RClass<P>.invoke(
         handler: RHandler<P>
     ): ReactElement =
@@ -69,30 +50,6 @@ interface RBuilder {
         children: List<Any> = emptyList()
     ): ReactElement =
         child(this, clone(props), children)
-
-    fun <P : RProps> child(
-        klazz: KClass<out Component<P, *>>,
-        handler: RHandler<P>
-    ): ReactElement =
-        klazz.rClass.invoke(handler)
-
-    fun <T, P : RProps> childFunction(
-        klazz: KClass<out Component<P, *>>,
-        handler: RHandler<P>,
-        children: RBuilder.(T) -> Unit
-    ): ReactElement =
-        child(
-            type = klazz.rClass,
-            props = RElementBuilder(jsObject<P>()).apply(handler).attrs,
-            children = listOf { value: T -> buildElement { children(value) } }
-        )
-
-    fun <P : RProps> node(
-        klazz: KClass<out Component<P, *>>,
-        props: P,
-        children: List<Any> = emptyList()
-    ): ReactElement =
-        klazz.rClass.node(props, children)
 
     fun RProps.children() {
         childList.addAll(Children.toArray(children))
@@ -164,23 +121,6 @@ interface RBuilder {
 @JsName("createBuilder")
 fun RBuilder(): RBuilder =
     RBuilderImpl()
-
-inline fun <P : RProps, reified C : Component<P, *>> RBuilder.child(
-    noinline handler: RHandler<P>
-): ReactElement =
-    child(C::class, handler)
-
-inline fun <T, P : RProps, reified C : Component<P, *>> RBuilder.childFunction(
-    noinline handler: RHandler<P>,
-    noinline children: RBuilder.(T) -> Unit
-): ReactElement =
-    childFunction(C::class, handler, children)
-
-inline fun <P : RProps, reified C : Component<P, *>> RBuilder.node(
-    props: P,
-    children: List<Any> = emptyList()
-): ReactElement =
-    node(C::class, props, children)
 
 open class RBuilderImpl : RBuilder {
     override val childList = mutableListOf<Any>()
@@ -273,32 +213,3 @@ fun <P : RProps> functionalComponent(
     return fc.unsafeCast<FunctionComponent<P>>()
 }
 
-/**
- * Append functional component [component] as child of current builder
- */
-@Suppress("EXTENSION_SHADOWED_BY_MEMBER")
-fun <P : RProps> RBuilder.child(
-    component: FC<P>,
-    props: P = jsObject(),
-    handler: RHandler<P> = {}
-): ReactElement =
-    child(component, props, handler)
-
-fun <T, P : RProps> RBuilder.childFunction(
-    component: FC<P>,
-    handler: RHandler<P> = {},
-    children: RBuilder.(T) -> Unit
-): ReactElement =
-    childFunction(component, jsObject<P>(), handler, children)
-
-fun <T, P : RProps> RBuilder.childFunction(
-    component: FC<P>,
-    props: P,
-    handler: RHandler<P> = {},
-    children: RBuilder.(T) -> Unit
-): ReactElement =
-    child(
-        type = component,
-        props = RElementBuilder(props).apply(handler).attrs,
-        children = listOf { value: T -> buildElement { children(value) } }
-    )
